pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Shiv <shivanshityagi@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.2.2"
python-dotenv = "^1.0.0"
web3 = "^6.0.0"
openai = "^0.27.2"

[tool.poetry.scripts]
bot = "bot:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
